+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - UIAppearance support.
+- turn haptic feedback off by setting `isHapticSnap` to `false`. 
 
 ## [1.9.1] - 2019-03-15